+++ apis/getSocialImage.ts
@@ -2,19 +2,17 @@ import buildQueryString from '@/utils/buildQueryString'
 import BASE_URL from './apiConfig'
 
 interface IGetSocialImageProps {
-  socialId: number
   accessToken: string
   imageFileExtension: 'JPEG' | 'JPG' | 'PNG' | 'WEBP'
 }
 
 async function getSocialImage({
-  socialId,
   accessToken,
   imageFileExtension,
 }: IGetSocialImageProps): Promise<Response> {
   const query = buildQueryString({ imageFileExtension })
   const response = await fetch(
-    `${BASE_URL}/socials/${socialId}/images${query ? `?${query}` : ''}`,
+    `${BASE_URL}/socials/images${query ? `?${query}` : ''}`,
     {
       method: 'GET',
       headers: {
