+++ components/GatheringCards/GatheringCardList.tsx
@@ -4,7 +4,7 @@ import { GatheringCard } from '@/components'
 import React from 'react'
 
 interface CardListProps {
-  data: GetSocialsType
+  socialsData: ISocials[]
 }
 
 /**
@@ -12,11 +12,11 @@ interface CardListProps {
  * @param data 데이터 배열
  */
 
-function GatheringCardList({ data }: CardListProps) {
+function GatheringCardList({ socialsData }: CardListProps) {
   return (
     <div className="grid w-full grid-cols-4 gap-x-20pxr gap-y-40pxr mb:grid-cols-2 mb:gap-x-10pxr mb:gap-y-32pxr tb:gap-x-16pxr max420:!grid-cols-1 max700Min480:!grid-cols-2 max900Min480:grid-cols-3">
-      {data.map((item, index) => (
-        <GatheringCard data={item} key={`${index + 0}`} />
+      {socialsData.map((item) => (
+        <GatheringCard data={item} key={item.id} />
       ))}
     </div>
   )
