+++ components/index.ts
@@ -20,7 +20,7 @@ export { default as CrewReviewWriteModal } from './CrewReviewWriteModal'
 export { default as PasswordFindForm } from './PasswordFindForm'
 export { default as Portal } from './Portal'
 export { default as InfoPortal } from './InfoPortal'
-export { default as SocialRegistrationForm } from './SocialRegistrationForm'
+export { default as SocialRegistrationForm } from './SocialRegistrationForm/SocialRegistrationForm'
 export { default as UploadImages } from './UploadImages/UploadImages'
 export { default as UploadImageCard } from './UploadImages/UploadImageCard'
 export { default as UploadImageCardsList } from './UploadImages/UploadImageCardsList'
